backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.2.1
python-multipart==0.0.6
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
